+++ Branch-SDK/build.gradle
@@ -6,13 +6,13 @@ def jarFileName
 dependencies {
     implementation fileTree(dir: 'libs', include: '*.jar')
     implementation 'androidx.annotation:annotation:1.1.0'
+    implementation 'com.android.installreferrer:installreferrer:1.1'
 
     // --- optional dependencies -----
     //Please note that the Branch SDK does not require any of the below optional dependencies to operate. This dependency is listed here so there will not be build errors,
     // but the library is *not* added to your app unless you do so yourself. Please check the code in gradle-mvn-push script to see how this works
 
     implementation 'com.google.firebase:firebase-appindexing:19.0.0'
-    implementation 'com.android.installreferrer:installreferrer:1.0'
 
     // testImplementation 'junit:junit:4.12'
     androidTestImplementation 'androidx.test.ext:junit:1.1.1'
